refactor: view collections abstraction
