+++ src/lib/components/ui/dialog/dialog-content.svelte
@@ -4,9 +4,10 @@
 	import { cn, flyAndScale } from '$lib/utils';
 	import { X } from 'lucide-svelte';
 
-	type $$Props = DialogPrimitive.ContentProps;
+	type $$Props = DialogPrimitive.ContentProps & { overlayClass?: HTMLElement['className'] };
 
 	let className: $$Props['class'] = undefined;
+	export let overlayClass: $$Props['overlayClass'] = undefined;
 	export let transition: $$Props['transition'] = flyAndScale;
 	export let transitionConfig: $$Props['transitionConfig'] = {
 		duration: 200
@@ -15,7 +16,7 @@
 </script>
 
 <Dialog.Portal>
-	<Dialog.Overlay />
+	<Dialog.Overlay class={cn('', overlayClass)} />
 	<DialogPrimitive.Content
 		{transition}
 		{transitionConfig}
