+++ src/lib/layouts/rest/views/view-collections/index.ts
@@ -1,0 +1,1 @@
+export { default as ViewCollections } from './view-collections.layout.svelte';
